feat: Add `created_at` to `workflow_run_jobs` schema (#275)
### Summary
This pull request adds the `created_at` property to the
`workflow_run_jobs` schema. This allows users to extract the creation
time of workflow run jobs and would be helpful in the tracking
individual jobs `queue_time`.

### Changes
- Updated `repository_streams.py` to include `created_at` in the
`schema` definition of the `workflow_run_jobs` stream.

### Testing
1. Confirmed that Github REST API endpoint for workflow jobs response
returns `created_at` field in the response.
2. Tested the changes locally using Meltano to ensure the `created_at`
field is correctly included and populated.