fix: Expand CollaboratorsStream primary keys (#278)
The collaborators stream pulls one row per selected repo per user who is
a known collaborator of that repo, so `id` (the user ID) alone isn't
sufficient to describe the grain. This PR adds `repo` and `org` to the
primary keys to capture the fact that there can be a row for each user x
repo pair.

My motivation is that I'm getting an error `Query error: UPDATE/MERGE
must match at most one source row for each target row ...` when trying
to extract this stream sequentially with `target-bigquery` and the
[upsert](https://hub.meltano.com/loaders/target-bigquery/#upsert-setting)
setting turned on, and that sounds related to the fact that the column
set as the primary key isn't distinctive.