fix: Tokens in config are no longer ignored when there are tokens in the environment, part of refactoring in preparation for app token refreshing (#284)
This PR continues the work started in
https://github.com/MeltanoLabs/tap-github/pull/281, completing the major
refactor. This is accomplished by creating PersonalTokenManager and
AppTokenManager classes, moving logic to them, and adding tests.

After this PR, I'll be able to easily implement refreshing for app
tokens and other app token related features.

In additional to the added unit tests, I have tested running an
extractor on top of the proposed code, in a couple scenarios:

with auth_token? | with additional_auth_tokens? | with personal tokens
in env? | with app token? | # tokens that should be used | # tokens used
in test run
--|--|--|--|--|--
no | no | no | yes | 1 | 1 ✅ 
yes | yes (1) | no | yes | 3 | 3 ✅ 
no | yes (2) | no | yes | 3 | 3 ✅ 

I am not easily able to test the case of personal tokens being detected
from the environment, just because changing the name of environment
variables is not simple in my set up.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>