feat: App token refreshing (#289)
This PR implements app token refreshing! Thank you for your patience
with the previous refactoring PRs that prepared for this one.

- As part of `has_calls_remaining`, the amount of time left until
estimated token expiry is checked, and tokens are refreshed if expiry is
close.
- A new config option `expiry_time_buffer` is added, controlling how
many minutes before expiry the app token will be refreshed. This
parallels how `rate_limit_buffer` is used to ensure we don't push tokens
all the way to their limits.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>