feat: Make ProjectItemsStream skip project if query can't be validated (#466)
We've seen this error thrown regularly:
```
singer_sdk.exceptions.FatalAPIError: (\"Graphql error: [{'message': 'Timeout on validation of query'}]\", <Response [200]>)
```

We suspect this happens because of project boards in the extraction
scope that have especially complex structures which lead to extra
complex GraphQL queries which are slower to validate. This PR makes it
so these projects will be skipped and a warning logged rather than
causing the tap to fail. I've tested this fix and it successfully
skipped problematic projects.