Tolerate 414 error in ProjectItemsStream
I've encountered a case where the ProjectItemsStream fails with this error:
```
414 Client Error: b'{"message": "We received a Request-URL that is too long from your client."}
```
Since ProjectItemsStream is constructing queries dynamically based on the project fields identified by the parent, it makes sense that it can construct a query so long the server rejects it.  I'm proposing to skip and move on in this case.